requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
Markdown==3.3.7
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
zipp==3.8.0
